+++ src/File/PortalNodeFilesystemStreamProtocolProvider.php
@@ -5,16 +5,17 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Bridge\ShopwarePlatform\File;
 
 use Heptacom\HeptaConnect\Core\Bridge\File\PortalNodeFilesystemStreamProtocolProviderInterface;
-use Heptacom\HeptaConnect\Core\Storage\Filesystem\FilesystemFactory;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\StorageKeyGeneratorContract;
+use League\Flysystem\FilesystemOperator;
 use M2MTech\FlysystemStreamWrapper\FlysystemStreamWrapper;
+use Shopware\Core\Framework\Adapter\Filesystem\PrefixFilesystem;
 
 final class PortalNodeFilesystemStreamProtocolProvider implements PortalNodeFilesystemStreamProtocolProviderInterface
 {
     public function __construct(
         private StorageKeyGeneratorContract $storageKeyGenerator,
-        private FilesystemFactory $filesystemFactory
+        private FilesystemOperator $filesystem
     ) {
     }
 
@@ -22,9 +23,11 @@ final class PortalNodeFilesystemStreamProtocolProvider implements PortalNodeFile
     {
         $key = $this->storageKeyGenerator->serialize($portalNodeKey);
         $streamScheme = \strtolower(\preg_replace('/[^a-zA-Z0-9]/', '-', 'hc-bridge-sw-' . $key));
+        $portalNodeId = $this->storageKeyGenerator->serialize($portalNodeKey->withoutAlias());
+        $normalizedPortalNodeId = \preg_replace('/[^a-zA-Z0-9]/', '_', $portalNodeId);
+        $filesystem = new PrefixFilesystem($this->filesystem, $normalizedPortalNodeId);
 
-        // TODO: Fix Flysystem compatibility
-        FlysystemStreamWrapper::register($streamScheme, $this->filesystemFactory->factory($portalNodeKey));
+        FlysystemStreamWrapper::register($streamScheme, $filesystem);
 
         return $streamScheme;
     }